Minor code readability refactor
